Implement single table diff and Export table diff to separate object and package
